Simplifies PyPI deployment over Travis configuration
Leverages the not so well documented token and `PYPI_PASSWORD` feature.
It's actually already possible in Travis v1 to use this environment
variable refs:
https://github.com/travis-ci/dpl/blob/v1.10.15/lib/dpl/provider/pypi.rb#L12
Note that the `__token__` user is a PyPI feature, refs:
https://pypi.org/help/#apitoken
This is more secure than user and password as tokens are generated with
uploaded only permissions and can be scoped to a given project.